Port triva everywhere
Would be better to pull this out as perhaps an extension method that you call instead of Accept.

There's a slight issue with nodes that are used as placeholders for later which are keys in a dictionary, but it might be possible to change the key mechanism to be tolerant of trivia being added anyway and would save writing this literally everywhere.